% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{accumulate}
\alias{accumulate}
\title{Perform a basin-to-basin flow accumulation}
\usage{
accumulate(basin_ids, downstream_ids, flows)
}
\arguments{
\item{basin_ids}{a vector of basin ids}

\item{downstream_ids}{a vector of downstream basin ids,
aligned with the entries in \code{basin_ids}}

\item{flows}{a vector of flows generated in each basin,
aligned with the entries in \code{basin_ids}}
}
\value{
a vector of outlet flows for each basin (including flow generated
        within the basin), aligned with the entries in \code{basin_ids}
}
\description{
Perform a basin-to-basin flow accumulation
}
