% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{accumulate_flow}
\alias{accumulate_flow}
\title{Accumulate flow, given flow directions and weights.}
\usage{
accumulate_flow(directions, weights, wrapX, wrapY)
}
\arguments{
\item{directions}{a matrix of flow directions, where directions are represented
by the following values:
\itemize{
\item east: 1
\item southeast: 2
\item south: 4
\item southwest: 8
\item west: 16
\item northwest: 32
\item north: 64
\item northeast: 128
\item none (sink cell): NA
}}

\item{weights}{a matrix of weights, representing the amount of flow originating at
each cell}

\item{wrapX}{should flow exiting the X-limits of the model be routed to the other side?}

\item{wrapY}{should flow exiting the Y-limits of the model be routed to the other side?}
}
\value{
a matrix of accumulated flow values
}
\description{
Accumulate flow, given flow directions and weights.
}
