% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_months.R
\name{add_months}
\alias{add_months}
\title{Add or subtract months from a given month}
\usage{
add_months(yyyymm, n)
}
\arguments{
\item{yyyymm}{Year/month in YYYYMM format}

\item{n}{Number of months to add}
}
\value{
year/month in YYYYMM format
}
\description{
Add or subtract months from a given month
}
