% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_flow_dirs.R
\name{adjust_flow_dirs}
\alias{adjust_flow_dirs}
\title{Adjust a flow direction matrix for compatibility with other layers}
\usage{
adjust_flow_dirs(fd, fd_extent, final_extent, final_dims)
}
\arguments{
\item{fd}{a flow direction matrix}

\item{fd_extent}{the spatial extent of \code{fd}, expressed as \code{xmin, xmax, ymin, ymax}}

\item{final_extent}{desired final extent, expressed as \code{xmin, xmax, ymin, ymax}}

\item{final_dims}{desired final dimensions}
}
\value{
adjusted version of \code{fd} as described above
}
\description{
Performs the following adjustments:
- if \code{fd} is at twice the resolution of other layers, but the latitude
  extents are not the same, moves one column of \code{fd} from left to right
  if doing so will make the extents equal
- pads \code{fd} with \code{NA} values for northern and southern latitudes
  included in \code{final_extent} but not \code{fd_extent}.
}
