% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{average_day_length}
\alias{average_day_length}
\title{Return the monthly average day length at given latitude(s)}
\usage{
average_day_length(latitudes, year, month)
}
\arguments{
\item{latitudes}{a vector of latitudes}

\item{year}{numeric year}

\item{month}{numeric month}
}
\value{
day length in hours
}
\description{
Return the monthly average day length at given latitude(s)
}
