% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_detained}
\alias{calc_detained}
\title{Calculate detained runoff and snowmelt}
\usage{
calc_detained(R, Pr, P, Sm, Dr, Ds, z, melt_month)
}
\arguments{
\item{R}{runoff}

\item{Pr}{precipitation}

\item{P}{net precipitation Pr - Sa + Sm}

\item{Sm}{snowmelt}

\item{Dr}{detained runoff}

\item{Ds}{detained snowmelt}

\item{z}{elevation}

\item{melt_month}{number of consecutive
months of melting conditions}
}
\description{
Calculate detained runoff and snowmelt
}
