% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_areas_m2.R
\name{cell_areas_m2}
\alias{cell_areas_m2}
\title{Compute the area occupied by each cell of a matrix}
\usage{
cell_areas_m2(extent, dim)
}
\arguments{
\item{extent}{a vector representing the spatial extent of the matrix
(\code{xmin, xmax, ymin, ymax})}

\item{dim}{the dimensions of the matrix (\code{nlat, nlon})}
}
\value{
a matrix having dimension \code{dim} and values representing
        the area occupied by each cell
}
\description{
Calculation is performed using spherical geometry
}
