% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{create_inward_dir_matrix}
\alias{create_inward_dir_matrix}
\title{For each pixel, compute which cells drain _into_ that pixel.}
\usage{
create_inward_dir_matrix(directions, wrapX, wrapY)
}
\arguments{
\item{directions}{a matrix of flow directions, where directions are represented
by the following values:
\itemize{
\item east: 1
\item southeast: 2
\item south: 4
\item southwest: 8
\item west: 16
\item northwest: 32
\item north: 64
\item northeast: 128
\item none (sink cell): NA
}}

\item{wrapX}{should flow exiting the X-limits of the model be routed to the other side?}

\item{wrapY}{should flow exiting the Y-limits of the model be routed to the other side?}
}
\value{
Matrix containing the summed direction values of all
        adjacent pixels that flow into this pixel. Value is
        zero if no adjacent pixels flow into this pixel (i.e,
        the pixel is a sink.)
}
\description{
For each pixel, compute which cells drain _into_ that pixel.
}
