% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{daily_hydro_loop}
\alias{daily_hydro_loop}
\title{Compute hydrological parameters for all pixels}
\usage{
daily_hydro_loop(P, Sa, Sm, E0, Ws, Wc, nDays, pWetDays)
}
\arguments{
\item{P}{precipitation for the time step [L]}

\item{Sa}{snow accumulation for the time step [L]}

\item{Sm}{snow melt for the time step [L]}

\item{E0}{potential evapotranspiration for the time step [L]}

\item{Ws}{soil moisture at start of time step [L]}

\item{Wc}{soil moisture holding capacity [L]}

\item{nDays}{number of days in time step [-]}

\item{pWetDays}{percentage of days with precipitation [-]}
}
\value{
a List of vectors of hydrological parameters:
          dWdt:   change in soil moisture [L]
          Ws_ave: average soil moisture over timestep [L],
          E:      evapotranspiration [L]
          R:      runoff [L]
}
\description{
Compute hydrological parameters for all pixels
}
