% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{day_length}
\alias{day_length}
\title{Return the day length at given latitude(s)}
\usage{
day_length(latitudes, year, month, day)
}
\arguments{
\item{latitudes}{a vector of latitudes}

\item{year}{numeric year}

\item{month}{numeric month}

\item{day}{numeric day of month}
}
\value{
day length in hours
}
\description{
Return the day length at given latitude(s)
}
