% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day_length_matrix.R
\name{day_length_matrix}
\alias{day_length_matrix}
\title{Compute a matrix of monthly average day length}
\usage{
day_length_matrix(year, month, extent, nrows, ncols)
}
\arguments{
\item{year}{year for computation. Must be >= 1900}

\item{month}{month for computation}

\item{extent}{vector of \code{(xmin, xmax, ymin, ymax)}}

\item{nrows}{number of rows (latitudes) in generated matrix}

\item{ncols}{number of columns (longitudes) in generated matrix}
}
\value{
a matrix of specified dimensions, where each cell represents
        the day length as a fraction of 24 hours
}
\description{
Compute a matrix of monthly average day length
}
