% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/days_in_yyyymm.R
\name{days_in_yyyymm}
\alias{days_in_yyyymm}
\title{Return the number of days in a given month}
\usage{
days_in_yyyymm(yyyymm)
}
\arguments{
\item{yyyymm}{Year/month in YYYYMM format}
}
\value{
number of days in month
}
\description{
Return the number of days in a given month
}
