% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{downstream_flow}
\alias{downstream_flow}
\title{Compute the sum of flow originating in downstream basins}
\usage{
downstream_flow(basin_ids, downstream_ids, flows)
}
\arguments{
\item{basin_ids}{a vector of basin ids}

\item{downstream_ids}{a vector of downstream basin ids,
aligned with the entries in \code{basin_ids}}

\item{flows}{a vector of flows generated in each basin,
aligned with the entries in \code{basin_ids}}
}
\value{
a vector of downstream flows for each basin (excluding flow generated
        within the basin), aligned with the entries in \code{basin_ids}
}
\description{
Compute the sum of flow originating in downstream basins
}
