% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doy_to_month.R
\name{doy_to_month}
\alias{doy_to_month}
\title{Return the month associated with a given day of year}
\usage{
doy_to_month(doy)
}
\arguments{
\item{doy}{day of year}
}
\value{
integer month associated with \code{doy}
}
\description{
Does not account for leap years
}
