% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_potential.R
\name{e_potential}
\alias{e_potential}
\title{Potential evapotranspiration using Hamon's equation}
\usage{
e_potential(lambda, T, nDays)
}
\arguments{
\item{lambda}{Average day length, as a fraction (0 to 1)}

\item{T}{Average temp (C)}

\item{nDays}{number of days to estimate}
}
\value{
Potential evapotranspiration (mm)
}
\description{
Potential evapotranspiration using Hamon's equation
}
