% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{g}
\alias{g}
\title{Unitless drying function}
\usage{
g(Ws, Wc, E0, P)
}
\arguments{
\item{Ws}{Soil moisture (mm)}

\item{Wc}{Soil water holding capacity (mm)}

\item{E0}{Potential evapotranspiration (mm/day)}

\item{P}{Effective precipitation (mm/day)}
}
\value{
Magnitude of decline in soil moisture (mm/day)
}
\description{
Unitless drying function
}
