% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_forcing.R
\name{is.wsim.lsm.forcing}
\alias{is.wsim.lsm.forcing}
\title{Determine if an object represents an LSM forcing}
\usage{
is.wsim.lsm.forcing(thing)
}
\arguments{
\item{thing}{object to test}
}
\description{
Determine if an object represents an LSM forcing
}
