% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_results.R
\name{is.wsim.lsm.results}
\alias{is.wsim.lsm.results}
\title{Determine if an object represents LSM model results}
\usage{
is.wsim.lsm.results(thing)
}
\arguments{
\item{thing}{object to test}
}
\description{
Determine if an object represents LSM model results
}
