% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_state.R
\name{is.wsim.lsm.state}
\alias{is.wsim.lsm.state}
\title{Determine if an object represents an LSM state}
\usage{
is.wsim.lsm.state(thing)
}
\arguments{
\item{thing}{object to test}
}
\description{
Determine if an object represents an LSM state
}
