% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{is_leap_year}
\alias{is_leap_year}
\title{Determine if a given year is a leap year}
\usage{
is_leap_year(y)
}
\arguments{
\item{y}{integer representation of year}
}
\value{
TRUE if the given year is a leap year
}
\description{
Determine if a given year is a leap year
}
