% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{make_daily_precip}
\alias{make_daily_precip}
\title{Compute precipitation for each day in a multi-day period}
\usage{
make_daily_precip(P_total, nDays, pWetDays)
}
\arguments{
\item{P_total}{total precipitation for time period}

\item{nDays}{number of days in time period}

\item{pWetDays}{fraction of days in which precipitation occurs}
}
\description{
Compute precipitation for each day in a multi-day period
}
