% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_results.R
\name{make_results}
\alias{make_results}
\title{Create a set of WSIM LSM results}
\usage{
make_results(
  Bt_RO,
  Bt_Runoff,
  E,
  EmPET,
  PET,
  PETmE,
  P_net,
  RO_m3,
  RO_mm,
  Runoff_m3,
  Runoff_mm,
  Sa,
  Sm,
  Ws,
  dWdt,
  extent,
  digits_mm,
  digits_m3
)
}
\arguments{
\item{Bt_RO}{accumulated total blue water runoff (taking account of detention) [m^3]}

\item{Bt_Runoff}{accumulated total blue water runoff (not taking account of detention) [m^3]}

\item{E}{evapotranspiration [mm]}

\item{EmPET}{actual minus potential evapotranspiration [mm]}

\item{PET}{potential evapotranspiration [mm]}

\item{PETmE}{potential minus actual evapotranspiration [mm]}

\item{P_net}{net precipitation [mm]}

\item{RO_m3}{runoff (taking account of detention) [m^3]}

\item{RO_mm}{runoff (taking account of detention) [mm]}

\item{Runoff_m3}{runoff (not taking account of detention) [m^3]}

\item{Runoff_mm}{runoff (not taking account of detention) [mm]}

\item{Sa}{snow accumulation [mm]}

\item{Sm}{snowmelt [mm]}

\item{Ws}{average soil moisture [mm]}

\item{dWdt}{change in soil moisture [mm]}

\item{extent}{spatial extent of input matrices \code{(xmin, xmax, ymin, ymax)}}

\item{digits_mm}{number of digits after decimal point to which parameters in
mm will be rounded}

\item{digits_m3}{number of digits after decimal point to which parameters in
m3 will be rounded}
}
\value{
\code{wsim.lsm.results} object containing supplied variables
}
\description{
Create a set of WSIM LSM results
}
