% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_state.R
\name{make_state}
\alias{make_state}
\title{Create a WSIM LSM state}
\usage{
make_state(extent, Snowpack, Dr, Ds, Ws, snowmelt_month, yearmon)
}
\arguments{
\item{extent}{spatial extent of input matrices \code{(xmin, xmax, ymin, ymax)}}

\item{Snowpack}{snowpack water equivalent [mm]}

\item{Dr}{detained runoff [mm]}

\item{Ds}{detained snowmelt [mm]}

\item{Ws}{soil moisture [mm]}

\item{snowmelt_month}{number of months of consecutive melting conditions}

\item{yearmon}{year and month of state (state should represent MM/01/YYYY)}
}
\value{
\code{wsim.lsm.state} object containing supplied variables
}
\description{
Create a WSIM LSM state
}
