% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{make_wet_day_list}
\alias{make_wet_day_list}
\title{Generate a list of days in which precipitation occurs}
\usage{
make_wet_day_list(nDays, pWetDays)
}
\arguments{
\item{nDays}{number of days in time period}

\item{pWetDays}{fraction of days in which precipitation falls}
}
\value{
A list of days in which precipitation falls, starting at 1
}
\description{
Generate a list of days in which precipitation occurs
}
