% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/next_yyyymm.R
\name{next_yyyymm}
\alias{next_yyyymm}
\title{Get the next month}
\usage{
next_yyyymm(yyyymm)
}
\arguments{
\item{yyyymm}{Year/month in YYYYMM format}
}
\value{
Following month in YYYYMM format
}
\description{
Get the next month
}
