% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_effective.R
\name{P_effective}
\alias{P_effective}
\title{Calculate the effective (net) precipitation}
\usage{
P_effective(Pr, Sa, Sm)
}
\arguments{
\item{Pr}{Measured precipitation (mm/month)}

\item{Sa}{Snow accumulation (mm/month)}

\item{Sm}{Snow melt (mm/month)}
}
\value{
Effective (net) precipitation
}
\description{
Calculate the effective (net) precipitation
}
