% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_forcing_from_cdf.R
\name{precip_mm}
\alias{precip_mm}
\title{Convert precipitation value to millimeters}
\usage{
precip_mm(precip, precip_units, duration, duration_units)
}
\arguments{
\item{precip}{precipitation values}

\item{precip_units}{units of \code{precip}}

\item{duration}{an optional duration over which precipitation falls}

\item{duration_units}{units of \code{duration}}
}
\value{
precipitation amount in millimeters
}
\description{
Convert precipitation value to millimeters
}
