% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_forcing_from_cdf.R
\name{read_forcing_from_cdf}
\alias{read_forcing_from_cdf}
\title{Read model forcing from a netCDF file}
\usage{
read_forcing_from_cdf(fname, yearmon)
}
\arguments{
\item{fname}{netCDF file containing forcing data}

\item{yearmon}{year and month associated with the forcing, in YYYYMM format}
}
\value{
\code{wsim.lsm.forcing} object containing model forcing
}
\description{
Read model forcing from a netCDF file
}
