% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_state_from_cdf.R
\name{read_state_from_cdf}
\alias{read_state_from_cdf}
\title{Read a model state from a netCDF file}
\usage{
read_state_from_cdf(fname)
}
\arguments{
\item{fname}{netCDF file containing state data}
}
\value{
\code{wsim.lsm.state} object containing model state.
}
\description{
Read a model state from a netCDF file
}
