% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Run a single iteration of the the model}
\usage{
run(static, state, forcing)
}
\arguments{
\item{static}{a list containing static inputs to the model.
The following variables are required:
\describe{
\item{elevation}{mean grid cell elevation in meters}
\item{flow_directions}{flow direction grid described in \link{accumulate_flow}}
\item{Wc}{soil moisture capacity in millimeters}
}}

\item{state}{a \code{wsim.lsm.state} object containing an input state for the model.
Described in \code{\link{make_state}}.}

\item{forcing}{a \code{wsim.lsm.forcing} object containing forcing for the model.
Described in \code{\link{make_forcing}}.}
}
\value{
a list containing model outputs and a state for the next time step:
             \describe{
             \item{obs}{a \code{wsim.lsm.results} object, as described in \code{\link{make_results}}}
             \item{next_state}{a \code{wsim.lsm.state} object containing the model state at the end of
                              the timestep. Described in \code{\link{make_state}}.}
             }
}
\description{
Run a single iteration of the the model
}
