% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saturated_vapor_pressure.R
\name{saturated_vapor_pressure}
\alias{saturated_vapor_pressure}
\title{Saturated vapor pressure using Buck's equation}
\usage{
saturated_vapor_pressure(Tm)
}
\arguments{
\item{Tm}{Daily mean temperature (C)}
}
\value{
Vapor pressure (kPa)
}
\description{
Saturated vapor pressure using Buck's equation
}
