% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{snow_accum}
\alias{snow_accum}
\title{Snow accumulation}
\usage{
snow_accum(Pr, T)
}
\arguments{
\item{Pr}{Measured precipitation (mm/day)}

\item{T}{Average daily temperature (C)}
}
\value{
snow accumulation in mm
}
\description{
Compute snow accumulation, assuming that all precipitation
is snowfall if the temperature less than -1 C, and that no
precipitation is snowfall if the temperature is unknown or
greater than or equal to -1 C.
}
