% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{snow_melt}
\alias{snow_melt}
\title{Calculate snowmelt}
\usage{
snow_melt(snowpack, melt_month, T, z)
}
\arguments{
\item{snowpack}{Snowpack (mm)}

\item{melt_month}{Number of consecutive months in which melting conditions
have been present}

\item{T}{Average temp (C)}

\item{z}{Elevation (m)}
}
\value{
Snowmelt (mm/month)
}
\description{
Melting occurs when the temperature is greater than -1 C.
If elevation is less than 500 m, all snow will melt in one timestep.
If elevation is greater than 500m, snowmelt will be divided over
two timesteps.
}
