% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{soil_moisture_change}
\alias{soil_moisture_change}
\title{Change in soil moisture}
\usage{
soil_moisture_change(P, E0, Ws, Wc)
}
\arguments{
\item{P}{Effective precipitation (mm/day)}

\item{E0}{Potential evapotranspiration (mm/day)}

\item{Ws}{Soil moisture (mm)}

\item{Wc}{Soil moisture holding capacity (mm)}
}
\value{
Change in soil moisture (mm/day)
}
\description{
Change in soil moisture
}
