% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_forcing_from_cdf.R
\name{temp_celsius}
\alias{temp_celsius}
\title{Convert temperature to degrees Celsius}
\usage{
temp_celsius(x, u)
}
\arguments{
\item{x}{temperature value}

\item{u}{units of \code{x}}
}
\value{
temperature value in degrees Celsius
}
\description{
Convert temperature to degrees Celsius
}
