% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_lsm_values_to_cdf.R
\name{write_lsm_values_to_cdf}
\alias{write_lsm_values_to_cdf}
\title{Write model values (state, observations, forcing) to netCDF file}
\usage{
write_lsm_values_to_cdf(values, fname, prec, attrs = NULL)
}
\arguments{
\item{values}{One of the following objects:
\itemize{
\item{\code{wsim.lsm.forcing}}
\item{\code{wsim.lsm.results}}
\item{\code{wsim.lsm.state}}
}}

\item{fname}{Output filename}

\item{prec}{optional variable-specific precision settings, as
described in \link{write_vars_to_cdf}}

\item{attrs}{optional attributes to write to output}
}
\description{
Write model values (state, observations, forcing) to netCDF file
}
